modify: modify all pro file

+++ Vibration_Peak_LSTM/workdir/Peak_LSTM_API.py
@@ -8,10 +8,16 @@ module_name = 'Peak_LSTM'
 
 # copy shap_func.py
 path_now = os.getcwd()
-# command_cp_1 = f'copy D:\\{module_name}\\workdir\\{module_name}.py '+ path_now  # Windows
-# command_cp_2 = f'copy D:\\{module_name}\\workdir\\utils.py '+ path_now  # Windows
-command_cp_1 = f'cp ../{module_name}.py '+ path_now  # Linux
-command_cp_2 = f'cp ../utils.py '+ path_now  # Linux
+import platform
+
+system_plat = platform.system()
+if system_plat == "Linux":
+    command_cp_1 = f'cp ../{module_name}.py '+ path_now  # Linux
+    command_cp_2 = f'cp ../utils.py '+ path_now  # Linux
+else:
+    command_cp_1 = f'copy D:\\{module_name}\\workdir\\{module_name}.py '+ path_now  # Windows
+    command_cp_2 = f'copy D:\\{module_name}\\workdir\\utils.py '+ path_now  # Windows
+
 os.system(command_cp_1)
 os.system(command_cp_2)
 
@@ -20,11 +26,13 @@ with open('log.txt', 'a') as fp:
     fp.write('running\n')
 
 os.system(command)
+if system_plat == "Linux":
+    os.system(f'rm {module_name}.py')  # Linux
+    os.system('rm utils.py')  # Linux
+else:
+    os.system(f'del {module_name}.py')  # Windows
+    os.system('del utils.py')  # Windows
 
-# os.system(f'del {module_name}.py')  # Windows
-os.system(f'rm {module_name}.py')  # Linux
-# os.system('del utils.py')  # Windows
-os.system('rm utils.py')  # Linux
 
 with tarfile.open('./result.tar.gz', 'w:gz') as tar:
     for filename in os.listdir():
